requirements.txt
customtkinter
tkinterdnd2
Pillow
cryptography
numpy
